+++ tests/endpoints/test_cargo_time_series.py
@@ -1,22 +1,22 @@
 from datetime import datetime
-from unittest import TestCase
 
+from tests.testcases import TestCaseUsingRealAPI
 from vortexasdk.endpoints.cargo_timeseries import CargoTimeSeries
 
 
-class TestCargoTimeSeries(TestCase):
+class TestCargoTimeSeries(TestCaseUsingRealAPI):
     def test_search_returns_all_days(self):
         start = datetime(2018, 10, 1)
         end = datetime(2019, 11, 1)
 
         df = (
             CargoTimeSeries()
-            .search(
+                .search(
                 filter_activity="loading_state",
                 filter_time_min=start,
                 filter_time_max=end,
             )
-            .to_df()
+                .to_df()
         )
 
         n_days = (end - start).days + 1